Include Gemfile.lock, make build more verbose for CI.
The Gemfile.lock is designed to be committed to version control:

  http://yehudakatz.com/2010/12/16/clarifying-the-roles-of-the-gemspec-and-gemfile/

It pins down dependency versions, and is only updated when a dev explicitly asks
for it to be. If it is not present, then the latest version of each dep
will be included each time a `bundle install` is run.

This relates to the build failure in #55. If I have jekyll-paginate-v2
pinned to v1.7.3 then it works. If I let it get chosen by `bundle
install` then we get v1.7.4 and it fails. Thus, the Gemfile.lock in this
commit includes v1.7.3.

As one final improvement, it also makes the build be more verbose so
that better output is shown when a failure does happen (--verbose
--trace).