Auto deploy feature branches to surge.sh.
This is an opt-in addition to the CI workflow, which enables the result
of the CI build to be pushed automatically to a public facing webserver.
Ideally this would be done via GitLab pages, but that doesn't yet allow
more than one branch to deploy. Instead, it uses surge.sh as a provider.

This does introduce a dependency on a new third party service. To ensure
this doesn't deter people from contributing, it is completely optional:
 * If you don't signup for surge, CI sill still pass
 * The new task `preview` is marked as `allow_failure: true`
 * If you do opt in and follow the instructions in CONTRIBUTING.md CI
will automatically:
   * Deploy the result of `jekyll build` to surge.sh
   * Link to this deployment from within the GitLab merge request UI
   * Teardown the deployment once the branch is removed from GitLab

Note that the regular deployment from `master` to GitLab pages remains
largely unchanged. This still deploys master to GitLab pages.

There are some other minor improvements in this commit also, such as:
 * Caching dependencies to make CI faster.
 * Commenting and formatting in .gitlab-ci.yml