Improve news to contain an excerpt etc.
This is mostly a migration of the old news page currently located at https://f-droid.org/news-and-reviews/.
The modified "posted on ... by ..." is now also used on posts' page.

I did not use `page.excerpt` because it produced inconsistent pieces of text.
Instead I used `page.content` and truncated it to 50 words.

Fixes #7.