Make surge.sh builds suffix with /fdroid-website too.
This brings them in line with what is deployed to gitlab pages.
When we move to deploying to the actual f-droid.org website, then we
should revisit whether this is a good idea. We want the surge.sh builds
to be as similar as possible to what gets deployed in the end, so
reviews of MRs can pick up any problems.