Only deploy GP repo to surge.sh
When deploying feature branches to surge.sh, don't try to deploy the
entire f-droid.org repo of apps. This results in about 80MiB for each
language we translate the website into, and quickly goes above the
100MiB limit on surge.sh.

This still gets deployed to GitLab pages if you push to master, and it
also builds the entire f-droid.org repo during the test stage to make
sure there are no bugs with it.