Add script to convert translated .po files back into .md files
The translated .po files which are pulled from weblate need to be
converted back into translated versions of the original .md files.

The process is as follows:

 * Copy the source .md file
 * Extract the `title:` attribute from the frontmatter
 * Strip the remaining frontmatter
 * Add a `# [TITLE]` line with the contents of the title attribute to
the destination .md file.

So far it was exactly as it was for the original conversion into .po
files. This is because the line numbers in the .po file need to match a
.md file. They can't match the original .md file because we stripped off
the frontmatter and added a new line to the start. Therefore, when
rebuilding i18n versions we need to use the same .md source that was
used to generate the po.

 * Use `po4a` to make an i18n version of the generated .md file.
 * Strip the first (title) line from this .md file
 * Add back the frontmatter from the original .md file
 * Replace the `title:` from the original frontmatter with the first
   line that was stripped from the translated .md file.
 * Add a `lang: ...` attribute to the new frontmatter.
 * Copy this .md file into an i18n'ed subdirectory of the original.
 * Delete all of the intermediate files when done.