keep deploy exclude list as part of this repository
This jekyll project does not generate all parts of the website. For example,
the repo itself is generated by fdroidserver.  The jekyll deploy procedure
should never be able to overwrite parts of the website that it does not
generate.  This rsync exclude list is all the parts of the website that
this jekyll project does not manage.