Ensure MultiViews is configured for all allowed languages.
The `update_langs.sh` script which updates the `_config.yml`
now also populates a `.supported-langs` file which is used
by the `prepare-multi-lang.sh` script when deciding which languages
should be supported. This could equally be done by traversing the
directories in the folder that Jekyll outputs (e.g. 'en/', 'fr/')
but once the script has run then it will no longer work.

This commit also makes a small change to allow the script to
be run as many times as you want. It will work the first time,
and each subsequent time it will recreate the symlinks again.