Fix CI, because it doesn't deploy into $OUT_DIR/$CI_PROJECT_NAME.
I mistakenly thought that it would deploy to $OUT_DIR/$CI_PROJECT_NAME
but it doesn't it just deploys to $OUT_DIR. The reference to
$CI_PROJECT_NAME is about the links that Jekyll puts inside the
generated HTML, not where it outputs to.

I think it only broke on master, hence why it passed before merging, but
failed once merged into master.