Add "artwork" repo as submodule and expose `/badge/*`.
The `/artwork` directory is now embedded in the site webroot.
This is a neccesity, because it needs to be in the top level directory
for Jekyll in order to get copied to the generated site.

However, we need `/badge` to be present, and contain the files inside
`/artwork/badge`. As such, we symlink from `/artwork/badge` -> `/badge`.

Ideally we'd do away with the `/artwork` directory completely and just
have the badges directory, but I'm not confident that it is possible
with git (i.e. cloning only a part of a submodule).