Only add artowrk in the f-droid.org CI task.
It is not required for other builds, and as such just ends up
taking up space and download time.

Given there is now important stuff in the website which is not
covered by CI automatically any more, I've added a "Production"
pipeline item to make it easier to discover the manually-runnable
"f-droid.org" pipeline job. The goal is to encourage release
taggers to run this task on master before tagging.