Also deploy legacy /forums for f-droid.org builds.
Uses the same technique as used for the /badge directory,
only this also updates that process to stream the .tar.gz file
rather than downloading the .zip file to disk.

Also, only copy to the built site when we are finished with Jekyll,
so that Jekyll doesn't need to waste time processing files which don't
need to be touched.