Remove QR from device screenshot, add below download.
Don't show it if on a mobile sized device, because you probably can't
take a picture of your own screen, and you may as well download it by
touching the link.

It is true that others can't take photos of your screen for the QR now,
but i imagine that is not a common use case.