docs: add how to create fdroid base.box from default debian vagrant boxes
The f-droid manual had no really good explanation on how to create a
buildserver base.box. It just refers to the vagrant docs which don't
have a very clear description either.

Also doing all those required steps manually is a time consuming and
error prone process. Especially when it needs to be redone for new
Debian releases.

So this commit documents an alternative way to prepare the base.box
relying on the debian prebuilt vagrant boxes. These are automatically
build with packer and signed by the Debian Vagrant team:
https://wiki.debian.org/Teams/Cloud/VagrantBaseBoxes

We still need to do a few adjustments on the Debian vagrant boxes,
namely install the virtualbox guest additions and provide a bigger hard
disk. We also disable the default rsync behavior of the box as this
conflicts with the fdroidserver cache directory mounting.

To do this we rely on two vagrant plugins and a small provisioning shell
script which enlarges the root partition and root filesystem.