Use slider rewind, instead of infinite scrolling. Seems less buggy.
The infinite scroll empirically resulted in several weird bugs, such as
not scrolling back to the start for certain apps ,and also when first
opening certain apps it would not allow you to scroll forwards at all.
The behaviour of the rewind option provides a similar UX (user can
continue scrolling forever) but seems a bit less buggy.