Added tutorials for swap/repomaker/adding repos.
Follows the designs provided by @crwinfrey in #112.

They are implemented as a collection called "tutorials". Any time a
new tutorial is added to the _tutorials directory then it will
automatically get put into the list of available tutorials and also will
be added to the paginator at the bottom of each tutorial.

Fixes #112.

Fixes repomaker#182.