mark pages with language code, and set direction for RtL languages
Right-to-Left languages like Arabic, Farsi, and Hebrew need to have the
direction of the text marked.  Browsers have an "auto" mode to auto-detect
the direction things should go.  Since we have mixed language pages, for
example app descriptions, we need to use auto mode, otherwise the English
will also be forced Right-to-Left.