enable caching of all files
This lets the browser cache all the files.  All files should be safely
cached, since they are all staticly generated.  I set the HTML to a short
cache time, since that's the most likely to change, and it is not very
large.  This like the index.json for the search should be cached longer,
but it should have a changing filename so that the cached version will be
instantly invalidated.