gitlab-ci: strip down to what actually works and is used
This config was so complicated, it couldn't make heads or tails of it. So
I'm reducing it to things that 100% work and are needed, and combining
the pages jobs (e.g. branches and master). This should make the whole thing
a lot more maintainable and understandable by anyone who comes along.

* all branches deploy to gitlab pages, easy for new contributors
* make test process the same on branches and master
* remove mostly broken surge.sh setup
* only limit languages on CI builds, not on the official deploy

fixes 2667bedf9f9dcf9651848ed18b5db7b52f7b9918