Make the tutorials overview page match the spec from carrie.
Some things I have left out of the spec, such as the blue heading. I've
left that to just be the same style as all the other headings for now.
Same with the link sto the tutorials, they are not underlined because
the rest of the site doesn't use underlined links.

Additionally, I had to add some extra frontmatter to the tutorials so
that they woudl be able to show appropriately in the overview.
Specifically, the heading on the overview page is different from the
heading of the tutorial. The tutorial is written as "How to ...",
whereas the main title shown in the tutorial list is more actionable,
selling features of F-Droid.