Don't use absolute URLs for image paths on this server.
In principle, we could probably try and avoid using `{{ site.url }}`
completely. The two places it is left in this commit is in the RSS feed
template, as the data inside a `<link> element [must begin with an
IANA-registered URI scheme](https://validator.w3.org/feed/docs/rss2.html),
and in the `<link type="canonical" ... />` because it seems very poorly
documented. Seems like you [can have relative paths in a canonical
link](https://webmasters.googleblog.com/2013/04/5-common-mistakes-with-relcanonical.html),
but who knows if that is still the case.

Fixes #171.