package: collapsible permission lists
This collapses the permission list for all versions. Having it repeated
for every version of an app is visually quite noisy and most of the time
doesn't add any information the user is interested in.

Instead we make all permission list collapsible and open the one from
the latest version by default.