standardize on italized field names, they are a "foreign language"
Italics are a standard way in many spoken languages for denoting words that
come from another language.  The YAML metadata field names are like another
language, so when field names are used within English text, they should be
italics rather than code blocks.  When there are used in snippets of code,
then they should be in a code block.