Workaround i18n bug in TWIF 47
Somewhere in the i18n process, the ampersand in the title in the YAML
header is substituted by the line itself, once. Like so:

title: "TWIF 46 & 47: Android SDK in Debian Buster and more website translations"

in the original becomes:

title: "TWIF 46 title: "TWIF 46 & 47: Android SDK in Debian Buster and more website translations" 47: Android SDK in Debian Buster and more website translations"

in any translated files.

This workaround substitutes the "&" for the word "and".