Cache-Control: immutable
Adding "immutable" means that the browser should not even check the ETag
if those files are cached.  That saves a network connection and response
round trip.  That should reduce the load on the webserver some, but only
really number of active TCP connections, rather than bandwidth, since the
files were already being cached.