overhaul donate options
Add a new /donate page, listing all our options and a preference for the
free platforms Liberapay and open collective.

This also makes it clear that all other donations go to F-Droid Ltd.

Overhaul the donate sidebar, remove everything except opencollective and
liberapay, add a 'more' link to the /donate page.

Finally add a redirect after downloading F-Droid to the donate page.
Thunderbird did a similar thing and dumped up their donations by an
order of magnitude or so. This should be fairly unobtrusive and it only
works with js enabled.