break out list of contributors into a data file with more details
Right now, the list of contributors is fed to the translators, which just
makes pointless work for them.  So I split out the contributor names into
a data file.  I then added the ability to also list names in other writing
systems, and include gitlab/weblate handles to link to.  Only a name is
required, the links and extra scripts are optional.