enable compression on HTML pages now that there is no more secrets
Compression on HTML pages can be abused to read info in the web page. Since
there are no longer any logins or passwords for the site, this has become
irrelevant.  All the content is public, and it is drastically easier to
tell which page a user is looking at by looking at the size of the file as
a unique fingerprint.