tools/set-website-posts-priority-in-weblate.py: prioritize in Weblate
By default, the posts are shown in Weblate based on the filename, which
means that the oldest posts get shown to translators first.  That is
backwards, so this sets the "priority" to make the newest posts the highest
priority.

Old TWIF posts are given a lower priority because they are quite a bit
harder to translate since they have lots of Markdown syntax.