verify our domains for the Google Search Console
Google gathers data on all websites that it sees, and makes analytics using
that data.  We can see what they have by "verifying" our site with them.
This merge request will set it up for a Google account that I have, I'm
happy to give others access there upon request.

It is important to note: this does not mean we are gathering any more data,
or making more data available to Google.  It just allows us to see the data
that Google is already gathering on us.