fix msgmerge warning
WARNING: Start of line didn't match any expected keyword.
WARNING: Problem on line 11030: | msgid "If, for whatever reason, you would simply prefer us not to distribute the app then we can only apologise. You deserve respect for your work which you've released under a Free Software license, and it wouldn't be pleasant for us to have to go against your wishes. However, the alternative is for us to deny use of the software to all our users."