👓 add note about using local `jekyll-fdroid` clone
The instrutions in the readme about using a local jekyll-fdroid clone
are not complete. When I tired to follow the instructions from the
readme I just eneded up with an error message I didn't understand.
(Something about specifying a branch, but no instructions where or how.)
So I figured it's way easier to just configure my local clone directly
in the Gemfile. This way I got a bit quicker too, because I don't have
to commit any change I want to try to git either.