🔨 fix locale issue in tools/make-release-tag.sh
The sort command in the old version of the script is interpreting
numbers based on locale. Our version strings (e.g. 2.179) are
interpreted as plain strings in some locales and as float numbers in
other locales. This makes it so that follosing sequence of releases will
have different results for the last release tag.

Lets assume this set of tags:

2.2, 2.179, 2.99

in some locales `sort` will produce an incorrect order:

2.179, 2.2, 2.99

where in other locales it'll get the correct order:

2.2, 2.99, 2.179

This change uses pythons packaging.version parser for ingesting git
tags, so we will get correct sorting results independant of locale.