replace only:changes: with rules:if:changes
spellcheckbot is triggered so that it can run in the fdroid group and get
access to the tokens it needs to post as @fdroid-bot. The current setup
leads to a bunch of unrelated jobs also being triggered:

https://gitlab.com/fdroid/fdroid-website/-/pipelines/1249478834

This stops that by converting only:changes: with rules:if:changes so that
triggers can be filtered out.