tools/prepare-for-deploy.sh: avoid git safe.directory errors
This script runs as root, and does git operations.  The fdroid-website dir
is all owned by fdroid:fdroid, so that triggers the git safe.directory
error.  This makes the script run the git commands as fdroid.

```
fatal: detected dubious ownership in repository at '/home/fdroid/fdroid-website'
To add an exception for this directory, call:

	git config --global --add safe.directory /home/fdroid/fdroid-website
```