Revert "Update Bitcoin and Litecoin URI to match client"
This reverts commit df54ed2adeb2a5a8553cfce582cecf3af7abea80.

* https://gitlab.com/fdroid/fdroid-website/-/merge_requests/1224?commit_id=e527934eb91c6644331f9c1006ec30d39e0e8499#note_2219752613
* https://github.com/linkchecker/linkchecker/pull/855

I don't think it really makes sense to use `litecoin:` URLs in the
webpage.  When people are browsing apps on the website, it means they
probably do not have the app installed.  So they are not likely to
donate for an app they have not installed.  The `litecoin:` URL is to
make a donation.  The link before shows information, and since the
user is viewing the app on the website, they are most likely looking
for information.

Also, df54ed2adeb2a5a8553cfce582cecf3af7abea80 breaks _linkchecker_:
https://gitlab.com/fdroid/fdroid-website/-/jobs/8420894335