robots.txt: include list of top AI bots, as reported by Cloudflare
https://blog.cloudflare.com/declaring-your-aindependence-block-ai-bots-scrapers-and-crawlers-with-a-single-click/

As part of our continued responsibility to provide a platform that puts
users in control, and respects developers, this adds to the list of AI
scaper bots that are disallowed by robots.txt.  This lets developers
decide themselves whether they want to feed their code and app texts to the
LLMs.  This follows up on !1056